requirements.txt
cryptography==0.9.3
django==1.8.3
djangorestframework==3.1.3
djangorestframework-jwt==1.6.0
pytz==2015.4